chore(tests/bpsec): merge  policies with tests
