+++ tests/bpsec/tests/pics_default.rs
@@ -2,7 +2,6 @@
 //!
 //! Tests RFC 9173 default security contexts against Hardy.
 
-use bpsec_tests::{PolicyAction, check_required_bcb, check_required_bib};
 use hardy_bpv7::block;
 use hardy_bpv7::bpsec::encryptor;
 use hardy_bpv7::bpsec::key;
@@ -11,6 +10,32 @@ use hardy_bpv7::bpsec::signer;
 use hardy_bpv7::bundle;
 use hardy_bpv7::cmp::compare_bundles;
 use hardy_bpv7::editor::{Chunk, Editor};
+
+#[derive(Debug, PartialEq, Eq)]
+pub enum PolicyAction {
+    Pass,
+    Reject,
+    RemoveBlock(u64),
+}
+
+pub fn check_required_bcb(bundle: &bundle::Bundle, target: u64) -> PolicyAction {
+    match bundle.blocks.get(&target) {
+        Some(blk) if blk.bcb.is_some() => PolicyAction::Pass,
+        Some(_) => PolicyAction::Reject,
+        None => PolicyAction::Reject,
+    }
+}
+
+pub fn check_required_bib(bundle: &bundle::Bundle, target: u64) -> PolicyAction {
+    match bundle.blocks.get(&target) {
+        Some(blk) if blk.bib != block::BibCoverage::None => PolicyAction::Pass,
+        Some(blk) if matches!(blk.block_type, block::Type::Payload | block::Type::Primary) => {
+            PolicyAction::Reject
+        }
+        Some(_) => PolicyAction::RemoveBlock(target),
+        None => PolicyAction::Reject,
+    }
+}
 
 fn integrity_key() -> key::Key {
     serde_json::from_value(serde_json::json!({
